chore: misc fixes/helpers for notebook

+++ tradezoo/agent/observation.py
@@ -35,7 +35,8 @@ class Observation:
                 self.stock_balance,
                 self.best_ask,
                 self.best_bid,
-            ]
+            ],
+            dtype=np.float32,
         )
